+++ src/pages/ClusterStatus.tsx
@@ -73,6 +73,7 @@ export class ClusterStatus extends BasePage {
             logs: this.generateLogsLink(),
             traces: this.generateTracesLink(),
             services: this.generateServicesLink(),
+            cost: this.generateCostLink(),
             edit: this.generateEditLink(),
             config: `/plugins/${APP_ID}`,
           }}