+++ src/pages/Events.tsx
@@ -87,6 +87,7 @@ export class EventsPage extends BasePage {
             logs: this.generateLogsLink(),
             traces: this.generateTracesLink(),
             services: this.generateServicesLink(),
+            cost: this.generateCostLink(),
             edit: this.generateEditLink(),
             config: `/plugins/${APP_ID}`,
           }}