+++ src/pages/NodesOverview.tsx
@@ -77,6 +77,7 @@ export class NodesOverviewPage extends BasePage {
             logs: this.generateLogsLink(),
             traces: this.generateTracesLink(),
             services: this.generateServicesLink(),
+            cost: this.generateCostLink(),
             edit: this.generateEditLink(),
             config: `/plugins/${APP_ID}`,
           }}